refactor(gateway): split into domain modules (MYM-16)
Decompose the single gateway.ts into the target module layout while keeping
one deployable service and identical route behavior:

  auth/      bearer.ts (the one shared bearerClaims token-verify seam +
             401/403 helpers), claims.ts (requireDocumentClaims scope guard)
  llm/       proxy.ts (Anthropic proxy), routes.ts (catch-all registration)
  documents/ routes.ts (scope-enforced /v1/documents/* handlers)
  db/        client.ts (Db seam/createDb), queries.ts (FTS/scope SQL)
  server.ts  createGateway composition root (was gateway.ts)
  env.ts     loadConfigFromEnv (was config.ts)

Token verification stays a single implementation (auth/bearer.ts), imported by
both the document guard and the LLM proxy. Document routes are still registered
before the LLM catch-all in server.ts (registration order is correctness-
critical). index.ts remains the Docker entrypoint.
